requirements.txt
requests
pandas
streamlit
plotly
python-dotenv